requirements.txt
altair
certifi
Flask

gunicorn
Jinja2
joblib
jupyter
numpy
MonsterLab
pandas
pymongo[srv]
python-dotenv
scikit-learn
scipy
